+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 
 - fix #2: only generate docs for classes and functions defined in the targeted module
+- fix #4: Class docs: Recognize properties that have a `help` attribute
 
 ### Fixed